refactor: migrate tests and update reply routes to use AgentManager
- Move all execution tests to dedicated test file (23 tests)
- Update reply.rs endpoints to use session-specific agents
- Make session_id optional for backward compatibility
- Auto-generate session ID if not provided
- Update confirm_permission and submit_tool_result handlers

This continues the session isolation work, ensuring each chat session
gets its own agent instance through the AgentManager.
