+++ crates/goose/src/execution/manager.rs
@@ -215,175 +215,3 @@ impl Default for AgentManager {
         Self::new()
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[tokio::test]
-    async fn test_session_isolation() {
-        let manager = AgentManager::new();
-
-        let session1 = SessionId::generate();
-        let session2 = SessionId::generate();
-
-        // Get agents for different sessions
-        let agent1 = manager
-            .get_agent(session1.clone(), ExecutionMode::chat())
-            .await
-            .unwrap();
-        let agent2 = manager
-            .get_agent(session2.clone(), ExecutionMode::chat())
-            .await
-            .unwrap();
-
-        // Should be different agents
-        assert!(!Arc::ptr_eq(&agent1, &agent2));
-
-        // Getting same session should return same agent
-        let agent1_again = manager
-            .get_agent(session1, ExecutionMode::chat())
-            .await
-            .unwrap();
-        assert!(Arc::ptr_eq(&agent1, &agent1_again));
-    }
-
-    #[tokio::test]
-    async fn test_execution_modes() {
-        let manager = AgentManager::new();
-
-        // Create agents with different modes
-        let interactive = manager
-            .get_agent(SessionId::generate(), ExecutionMode::Interactive)
-            .await
-            .unwrap();
-
-        let background = manager
-            .get_agent(SessionId::generate(), ExecutionMode::Background)
-            .await
-            .unwrap();
-
-        let subtask = manager
-            .get_agent(
-                SessionId::generate(),
-                ExecutionMode::SubTask {
-                    parent_session: "parent-123".to_string(),
-                },
-            )
-            .await
-            .unwrap();
-
-        // All should be different agents
-        assert!(!Arc::ptr_eq(&interactive, &background));
-        assert!(!Arc::ptr_eq(&background, &subtask));
-        assert!(!Arc::ptr_eq(&interactive, &subtask));
-    }
-
-    #[tokio::test]
-    async fn test_session_limit() {
-        let manager = AgentManager::with_max_sessions(3);
-
-        // Create 3 sessions
-        let sessions: Vec<_> = (0..3)
-            .map(|i| SessionId::from(format!("session-{}", i)))
-            .collect();
-
-        for session in &sessions {
-            manager
-                .get_agent(session.clone(), ExecutionMode::chat())
-                .await
-                .unwrap();
-        }
-
-        assert_eq!(manager.session_count().await, 3);
-
-        // Creating 4th should evict oldest
-        let new_session = SessionId::from("session-new");
-        manager
-            .get_agent(new_session, ExecutionMode::chat())
-            .await
-            .unwrap();
-
-        // Should still have only 3 sessions
-        assert_eq!(manager.session_count().await, 3);
-
-        // First session should have been evicted
-        assert!(!manager.has_session(&sessions[0]).await);
-    }
-
-    #[tokio::test]
-    async fn test_remove_session() {
-        let manager = AgentManager::new();
-        let session = SessionId::from("remove-test");
-
-        // Create session
-        manager
-            .get_agent(session.clone(), ExecutionMode::chat())
-            .await
-            .unwrap();
-        assert!(manager.has_session(&session).await);
-
-        // Remove it
-        manager.remove_session(&session).await.unwrap();
-        assert!(!manager.has_session(&session).await);
-
-        // Removing again should error
-        assert!(manager.remove_session(&session).await.is_err());
-    }
-
-    #[tokio::test]
-    async fn test_concurrent_access() {
-        use std::sync::Arc;
-
-        let manager = Arc::new(AgentManager::new());
-        let session = SessionId::from("concurrent-test");
-
-        // Spawn multiple tasks accessing the same session
-        let mut handles = vec![];
-        for _ in 0..10 {
-            let mgr = Arc::clone(&manager);
-            let sess = session.clone();
-            let handle =
-                tokio::spawn(
-                    async move { mgr.get_agent(sess, ExecutionMode::chat()).await.unwrap() },
-                );
-            handles.push(handle);
-        }
-
-        // Collect all agents
-        let agents: Vec<_> = futures::future::join_all(handles)
-            .await
-            .into_iter()
-            .map(|r| r.unwrap())
-            .collect();
-
-        // All should be the same agent
-        for agent in &agents[1..] {
-            assert!(Arc::ptr_eq(&agents[0], agent));
-        }
-
-        // Only one session should exist
-        assert_eq!(manager.session_count().await, 1);
-    }
-
-    #[tokio::test]
-    async fn test_execute_recipe_stub() {
-        let manager = AgentManager::new();
-        let session = SessionId::generate();
-
-        let recipe = serde_json::json!({
-            "name": "test_recipe",
-            "instructions": "test"
-        });
-
-        let result = manager
-            .execute_recipe(session.clone(), recipe, ExecutionMode::Interactive)
-            .await
-            .unwrap();
-
-        // Verify stub response
-        assert_eq!(result["status"], "ready_for_future");
-        assert_eq!(result["agent_created"], true);
-        assert_eq!(result["session_id"], session.to_string());
-    }
-}
