+++ crates/goose/src/execution/mod.rs
@@ -95,58 +95,3 @@ impl From<&str> for SessionId {
         Self(s.to_string())
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_execution_mode_helpers() {
-        assert_eq!(ExecutionMode::chat(), ExecutionMode::Interactive);
-        assert_eq!(ExecutionMode::scheduled(), ExecutionMode::Background);
-
-        let parent = "parent-123".to_string();
-        assert_eq!(
-            ExecutionMode::task(parent.clone()),
-            ExecutionMode::SubTask {
-                parent_session: parent
-            }
-        );
-    }
-
-    #[test]
-    fn test_session_id_generation() {
-        let id1 = SessionId::generate();
-        let id2 = SessionId::generate();
-
-        // Should be unique
-        assert_ne!(id1, id2);
-
-        // Should be valid UUIDs
-        assert_eq!(id1.0.len(), 36); // UUID string length
-        assert_eq!(id2.0.len(), 36);
-    }
-
-    #[test]
-    fn test_session_id_from_string() {
-        let id_str = "test-session-123";
-        let session_id = SessionId::from_string(id_str.to_string());
-        assert_eq!(session_id.as_str(), id_str);
-
-        // Test From trait
-        let session_id2: SessionId = id_str.into();
-        assert_eq!(session_id, session_id2);
-    }
-
-    #[test]
-    fn test_display_traits() {
-        let session = SessionId::from_string("display-test".to_string());
-        assert_eq!(format!("{}", session), "display-test");
-
-        let mode = ExecutionMode::Interactive;
-        assert_eq!(format!("{}", mode), "interactive");
-
-        let mode2 = ExecutionMode::task("parent-456".to_string());
-        assert_eq!(format!("{}", mode2), "subtask(parent: parent-456)");
-    }
-}
